perf: pool render buffers, fix rasterizer O(N^2) taper and dab bugs
Production hardening of the CPU render path:

- Add PixelBufferPool (bounded, thread-safe) and route Compositor
  composite/renderLayerContent through it, replacing per-pass
  canvas-sized allocations that caused GC pauses and OOM crashes.
- StrokeRasterizer: thread accumulated distance through the dab loop so
  taper is O(1) per dab (was O(N) via indexOf, O(N^2) per stroke);
  remove the dead `segmentT * 0f` term; stamp the path-end dab once
  after the whole stroke instead of once per segment, eliminating
  double-darkening at segment joints.
- Enforce the selection mask at the dab/capsule primitive so strokes
  cannot bleed outside a selection before post-compositing.
- Make the rasterizer scratch buffer ThreadLocal and lastDabCount
  volatile (AtomicInteger) for GL-thread/worker concurrency.
- Fix vector-layer thumbnails: size them from the strokes' bounding box
  (+ padding) instead of passing 0/1 dimensions, and return null for
  empty layers.
- Build hygiene: migrate Hilt/Room from kapt to KSP, tighten Hilt
  ProGuard keeps so R8 can shrink the DI runtime, declare the required
  OpenGL ES 2.0 feature in the manifest, use double division for the
  GB storage summary, and drop stale detekt baseline entries.

+++ app/src/main/java/com/artflow/studio/presentation/ui/viewmodel/MainViewModel.kt
@@ -262,10 +262,15 @@ class MainViewModel
 
         fun storageSummary(): String {
             val bytes = storage.totalStorageBytes()
+            // Double division: Float mantissa precision loses visible accuracy once the value
+            // exceeds 16 MB of bytes.
             return when {
                 bytes < 1024 * 1024 -> "%.0f KB on device".format(bytes / 1024f)
                 bytes < 1024L * 1024 * 1024 -> "%.1f MB on device".format(bytes / (1024f * 1024f))
-                else -> "%.2f GB on device".format(bytes / (1024f * 1024f * 1024f))
+                else ->
+                    "%.2f GB on device".format(
+                        bytes.toDouble() / (1024.0 * 1024.0 * 1024.0),
+                    )
             }
         }
     }
